+++ src/components/content/Links.vue
@@ -199,9 +199,7 @@
 </template>
 
 <script>
-import { computed } from 'vue';
-
-import { useContentBaseComponent } from '@/components/content/contentBase';
+import { useAppLinksComponent } from '@/components/content/appLinks';
 
 export default {
 	name: 'Links',
@@ -216,70 +214,19 @@ export default {
 			noBreakingSpaces,
 			notImplementedError,
 			success,
-			serviceStore
-		} = useContentBaseComponent(props, context);
-
-		const links = computed(() => {
-			let temp = serviceStore.getters.getContent();
-			if (!temp)
-				return [];
-			if (!temp.links)
-				return [];
-			return temp.links.filter(l => l.enabled);
-		});
-		const linksBooks = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'book');
-		});
-		const linksCompetitions = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'competition');
-		});
-		const linksEvents = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'event');
-		});
-		const linksForums = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'forum');
-		});
-		const linksGeneral = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => String.isNullOrEmpty(l.category));
-		});
-		const linksGuidance = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'guidance');
-		});
-		const linksManufacturers = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'manufacturer');
-		});
-		const linksOrganizations = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'organization');
-		});
-		const linksVendors = computed(() => {
-			if (!links.value)
-				return [];
-			return links.value.filter(l => l.category === 'vendor');
-		});
-		
-		const target = (item) => {
-			return !item.local ? 'blank' : '';
-		};
-
-		const sort = (links) => {
-			links.sort((a, b) => a.order >= b.order);
-		};
+			links,
+			linksBooks,
+			linksCompetitions,
+			linksEvents,
+			linksForums,
+			linksGeneral,
+			linksGuidance,
+			linksManufacturers,
+			linksOrganizations,
+			linksVendors,
+			sort,
+			target
+		} = useAppLinksComponent(props, context);
 
 		return {
 			correlationId,
